refactor: refactor button styles
