--- src/app/components/button-outlined/button-outlined.scss
@@ -1,29 +1,0 @@
-@import 'Styles/constants';
-@import 'Styles/placeholders';
-@import 'Styles/mixins';
-
-.button-outlined {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  padding: 10px 24px;
-  margin: 24px 0;
-  width: max-content;
-  font-size: 18px;
-  font-weight: $font-weight-medium;
-  color: $color-primary-black;
-  border: 2px solid currentColor;
-  @extend %transition-all;
-
-  & .arrow {
-    position: relative;
-    top: 2px;
-    margin-left: 8px;
-  }
-
-  @include media-hover {
-    &:hover {
-      background-color: $color-secondary-gray100;
-    }
-  }
-}
