--- src/app/components/button-outlined/button-outlined.test.tsx
@@ -1,24 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import { describe, expect, it } from 'vitest';
-import { ButtonOutlined } from './button-outlined';
-
-describe('ButtonOutlined', () => {
-  it('should render correctly', () => {
-    render(<ButtonOutlined label="Test" href="http://test.com" />);
-    const button = screen.getByRole('link', { name: /Test/i });
-    expect(button).toBeInTheDocument();
-  });
-
-  it('should have correct label and href properties', () => {
-    render(<ButtonOutlined label="Test" href="http://test.com" />);
-    const button = screen.getByRole('link', { name: /Test/i });
-    expect(button).toHaveAttribute('href', 'http://test.com');
-    expect(button).toHaveTextContent('Test');
-  });
-
-  it('should display an arrow symbol', () => {
-    render(<ButtonOutlined label="Test" href="http://test.com" />);
-    const arrow = document.getElementById('btn-arrow');
-    expect(arrow).toBeInTheDocument();
-  });
-});
