--- src/app/components/button-outlined/button-outlined.tsx
@@ -1,5 +1,0 @@
-import { MakeBtnTemp } from '../button/button';
-
-import './button-outlined.scss';
-
-export const ButtonOutlined = MakeBtnTemp(true);
