--- src/app/components/button-outlined/index.ts
@@ -1,1 +1,0 @@
-export { ButtonOutlined } from './button-outlined';
