+++ src/app/components/button/button.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import { Button } from './button';
+import { Button, ButtonOutlined } from './button';
 
 describe('Button', () => {
   const label = 'Click me';
@@ -28,5 +28,26 @@ describe('Button', () => {
     expect(link).toHaveAttribute('rel', 'noreferrer');
     expect(link).toHaveTextContent('Click me');
     expect(link).toHaveAttribute('target', '_blank');
+  });
+});
+
+describe('ButtonOutlined', () => {
+  it('should render correctly', () => {
+    render(<ButtonOutlined label="Test" href="http://test.com" />);
+    const button = screen.getByRole('link', { name: /Test/i });
+    expect(button).toBeInTheDocument();
+  });
+
+  it('should have correct label and href properties', () => {
+    render(<ButtonOutlined label="Test" href="http://test.com" />);
+    const button = screen.getByRole('link', { name: /Test/i });
+    expect(button).toHaveAttribute('href', 'http://test.com');
+    expect(button).toHaveTextContent('Test');
+  });
+
+  it('should display an arrow symbol', () => {
+    render(<ButtonOutlined label="Test" href="http://test.com" />);
+    const arrow = document.getElementById('btn-arrow');
+    expect(arrow).toBeInTheDocument();
   });
 });
