+++ src/app/components/button/index.ts
@@ -1,1 +1,1 @@
-export { Button } from './button';
+export { Button, ButtonOutlined } from './button';
