+++ src/app/components/index.ts
@@ -2,7 +2,6 @@ export { PrincipleCard, type PrincipleCardProps } from './principle-card';
 export { SocialMedia, type SocialMediaProps } from './social-media';
 export { OptionItem, type OptionItemProps } from './option-item';
 export { EventCard, type EventCardProps } from './event-card';
-export { ButtonOutlined } from './button-outlined';
 export { SectionLabel } from './section-label';
 export { Breadcrumbs } from './breadcrumbs';
 export { Title, TitleType } from './title';
@@ -12,7 +11,7 @@ export { Paragraph } from './paragraph';
 export { DateLang } from './date-lang';
 export { Subtitle } from './subtitle';
 export { Actions } from './actions';
-export { Button } from './button';
+export { Button, ButtonOutlined } from './button';
 export { Places } from './places';
 export { MobileView } from './mobile-view';
 export { LogoWrapper } from './logo-wrapper';
