+++ src/features/home/numbers/numbers.scss
@@ -134,7 +134,7 @@
   .info-divider {
     width: 3px;
     height: 180px;
-    background-color: #e2e2e2;
+    background-color: $color-secondary-gray180;
     @include media-tablet {
       height: 140px;
     }
