+++ src/styles/_constants.scss
@@ -19,10 +19,14 @@ $color-secondary-gray600: #545454;
 $color-secondary-gray500: #757575;
 $color-secondary-gray400: #afafaf;
 $color-secondary-gray200: #e6e6e6;
+$color-secondary-gray180: #e2e2e2;
+$color-secondary-gray150: #eeeeee;
 $color-secondary-gray100: #f2f2f2;
 $color-secondary-brown: #664d1b;
 $color-secondary-yellow: #ffdb20;
 $color-secondary-yellow-50: #fffaf0;
+$color-secondary-white-50: #fefefe;
 $color-secondary-red: #d44333;
+$color-secondary-blue-100: #eef3fe;
 
 $color-accent-yellow: rgba(255, 219, 32, 0.8);
